backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "ashleyk9318",email = "ashleyk9318@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cassandra-driver (>=3.29.2,<4.0.0)",
    "flask (>=3.1.1,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
